Add Calendar.Settings Class
This class replaces the old Maya.Settings namespace and provide access
to internal and external properties even when sandboxed.

For internal settings, this will use the GLib.Settings if disponible.
For external settings, this will first try to use the portal,
fallbacking to the GLib.Settings if it fails for any reason.