GridDay: Fix day in month in grid with GLib 2.73.1+
GLib 2.73.1 no longer fires notify signal for properties set during instantiation:
https://gitlab.gnome.org/GNOME/glib/-/commit/8851112b8de024ff6fb4235a4475ffa6c98973dc
This resulted in the label remaining empty.

Now, the `Grid.update_day` would also run and set `date` property
but that did not have an effect – probably because the date did not change.

We need to set the initial label ourselves just like we do it in notify signal.
To avoid code duplication, we can just use `GBinding`.

Co-authored-by: Corentin Noël <tintou@noel.tf>